Changed migration to remove index on :code in fm_rates whatever the name of the index might be

+++ db/migrate/20200113193447_change_fm_rates_attributes.rb
@@ -3,7 +3,8 @@ class ChangeFmRatesAttributes < ActiveRecord::Migration[5.2]
     add_column :fm_rates, :standard, :string, limit: 1
     add_column :fm_rates, :direct_award, :boolean
     change_column :fm_rates, :code, :string, unique: false, limit: 5
-    remove_index :fm_rates, ['code']
+    remove_index :fm_rates, ['code'] if index_exists?(:fm_rates, ['code'])
+    remove_index :fm_rates, name: 'fm_rates_code_key' if index_exists?(:fm_rates, name: 'fm_rates_code_key')
     add_index :fm_rates, ['code'], unique: false
   end
